requirements.txt
Flask==3.1.0
Pillow==12.0.0
pypdfium2==4.30.0
gunicorn==23.0.0